+++ src/jobs/ImportJob.php
@@ -45,9 +45,6 @@ class ImportJob extends BaseBatchedJob
             return;
         }
 
-        // Call for max power
-        Campaign::$plugin->maxPowerLieutenant();
-
         parent::execute($queue);
 
         // TODO: move what’s below this into the `BaseBatchedJob::after` method in Campaign 3.